backend/requirements.txt
Django==3.2.18
celery[redis]==5.2.7
djoser==2.1.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
gunicorn==20.0.4
psycopg2-binary==2.9.3
python-dotenv==1.0.0
reportlab==3.6.12
python-telegram-bot==20.7
